HTTPMonitoringInterceptor.java

+++ src/main/java/works/weave/socks/cart/middleware/HTTPMonitoringInterceptor.java
@@ -32,7 +32,7 @@ public class HTTPMonitoringInterceptor implements HandlerInterceptor {
             httpServletResponse, Object o, ModelAndView modelAndView) throws Exception {
         long start = (long) httpServletRequest.getAttribute(startTimeKey);
         long elapsed = System.nanoTime() - start;
-        double seconds = (double) elapsed / 1000000000.0;
+        double seconds = elapsed / 1000000000.0;
         requestLatency.labels(
                 serviceName,
                 httpServletRequest.getMethod(),
